shellcheck fixes

+++ usr/lib/archboot/iso.sh
@@ -174,7 +174,7 @@ _prepare_uefi_image() {
     mkfs.vfat --invariant -C "${VFAT_IMAGE}" "${IMGSZ}" >/dev/null
     ## Copying all files to UEFI vfat image
     mcopy -m -i "${VFAT_IMAGE}" -s "${_ISODIR}"/EFI "${_ISODIR}"/boot ::/
-    rm -r "${_ISODIR}"/EFI "${_ISODIR}"/boot
+    rm -r "${_ISODIR}"/EFI "${_ISODIR:?}"/boot
 }
 
 _prepare_extlinux_conf() {
